Change configuration file directory
* change configuration file directory: `src/configs/files` > `resources/configs`
* configuration file loading mechnics refined
  * every files under `resources`
  * resource files under `resources/locale`
  * configuration files under `resources/configs`
