+++ src/configs/modules/readConfigFile.ts
@@ -1,5 +1,6 @@
 import IConfiguration from '#configs/interfaces/IConfiguration';
 import logging from '#logger/bootstrap';
+import getCwd from '#tools/misc/getCwd';
 import fs from 'fs';
 import { parse } from 'jsonc-parser';
 import path from 'path';
@@ -7,7 +8,7 @@ import path from 'path';
 const log = logging(__filename);
 
 export default function readConfigFile(runMode: string): IConfiguration {
-  const dirname = path.join(__dirname, '..', 'files');
+  const dirname = path.join(getCwd(process.env), 'resources', 'configs');
   const filename = `config.${runMode}.json`;
   const configBuf = fs.readFileSync(path.join(dirname, filename));
   const parsed = parse(configBuf.toString()) as IConfiguration;
