+++ src/configs/modules/readJsonSchemaFile.ts
@@ -1,5 +1,6 @@
 import type ISchemaDatabaseItem from '#configs/interfaces/ISchemaDatabaseItem';
 import logging from '#logger/bootstrap';
+import getCwd from '#tools/misc/getCwd';
 import fs from 'fs';
 import { parse } from 'jsonc-parser';
 import path from 'path';
@@ -7,7 +8,7 @@ import path from 'path';
 const log = logging(__filename);
 
 export default function readJsonSchemaFile(): Record<string, ISchemaDatabaseItem> {
-  const dirname = path.join(__dirname, '..', 'files');
+  const dirname = path.join(getCwd(process.env), 'resources', 'configs');
   const filename = 'store.json';
   const jsonSchemaBuf = fs.readFileSync(path.join(dirname, filename));
   const parsed = parse(jsonSchemaBuf.toString()) as Record<string, ISchemaDatabaseItem>;
