+++ src/handlers/get/v1/poke-detail/[name].ts
@@ -10,7 +10,8 @@ import { FastifyRequest, RouteShorthandOptions } from 'fastify';
 export const option: RouteShorthandOptions = {
   schema: {
     tags: ['Pokemon'],
-    summary: 'Pokemon detail information using by name',
+    summary: 'Pokemon detail',
+    description: 'Pokemon detail information using by name',
     querystring: { $ref: 'IReqPokeDetailQuerystring' },
     params: { $ref: 'IReqPokeDetailParams' },
     response: {
@@ -25,6 +26,6 @@ export default async function readPokeDetailByNameHandler(
   req: FastifyRequest<{ Querystring: IReqPokeDetailQuerystring; Params: IReqPokeDetailParams }>,
 ) {
   const resp = await readPokeDetailByName(req.params.name);
-  const serialized = iPokemonDto.transformPokemonToWithTid(resp.data, req.query.tid);
+  const serialized = iPokemonDto.fromTid(resp.data, req.query.tid);
   return serialized;
 }
