+++ src/transforms/v1/IPokemonDto.ts
@@ -1,12 +1,9 @@
 import IPokemonDto from '#dto/v1/poke-detail/IPokemonDto';
 
-function transformPokemonToWithTid(
-  pokemon: IPokemonDto,
-  tid: string,
-): IPokemonDto & { tid: string } {
+function fromTid(pokemon: IPokemonDto, tid: string): IPokemonDto & { tid: string } {
   return { ...pokemon, tid };
 }
 
 export default {
-  transformPokemonToWithTid,
+  fromTid,
 };
